requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
google-generativeai>=0.8.0
anthropic>=0.18.1
imaplib2==3.6
selenium==4.17.2
pyautogui==0.9.54
requests==2.31.0
pydantic==2.6.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Redis & Caching
redis>=5.0.0
hiredis>=2.3.0

# Background Tasks
celery>=5.3.0
